connectivity: use the right test namespace in cnps
Make sure that we install CPNs into the right namespace, as well as that
any rules reference this namespace. A placeholder had to be picked (or
otherwise all rules iterated, which is harder to get right) so I am
picking an envsubst-compatible one just in case we ever need to pipe
these into 'kubectl' for testing.

Signed-off-by: Bruno Miguel Custódio <brunomcustodio@gmail.com>